+++ src/squirrel/squirrelMeasure.h
@@ -42,7 +42,7 @@ public:
     bool isValid() { return valid; }
     QString Error() { return err; }
     qint64 GetObjectID() { return objectID; }
-    void SetObjectID(int id) { objectID = id; }
+    void SetObjectID(qint64 id) { objectID = id; }
 
     qint64 subjectRowID;
